+++ gazeplay-commons/src/main/java/net/gazeplay/commons/configuration/observableproperties/PropertiesDoubleProperty.java
@@ -1,4 +1,4 @@
-package net.gazeplay.commons.configuration;
+package net.gazeplay.commons.configuration.observableproperties;
 
 import javafx.beans.property.DoublePropertyBase;
 import lombok.RequiredArgsConstructor;
